Move external library requests to Cloudflare
The gtmetrix.com site complains that the demo shop uses more than four domains. This change moves the jQuery, BootStrap, and FontAwesome scripts to the CDNJS Cloudflare CDN. That drops the number of domains on the demo site to exactly four: template.me.uk, cdnjs.cloudflare.com, www.googletagmanager.com, and www.google-analytics.com. The Google site links are not distributed with the core install and probably need to be downloaded directly regardless. So this leaves only the store itself and the CDNJS domain for the core install.

* Migrate from the fontawesome host to Cloudflare.

* Switch from JQuery.com to CDNJS.CloudFlare.

* Use CDNJS Cloudflare to serve Bootstrap.