Simplify category tree
This simplifies the logic for extracting information from the
category tree.

Primarily shifts from using foreach/if key equal logic to isset
key.  This pattern was used multiple times but is unnecessary.

Note that where there were nested foreach loops, it often retains
one.  Because there was often a need for one but not two.

Also renames getChildren to getDescendants, as it recursively gets
the children of the categories it adds.  And makes various other
formatting and logic changes.