Switch from mysqli_ functions to OO methods
PHP is slowly transitioning from a procedural model to an object-
oriented model for database handling.  This moves forward to match
those changes.

It modifies the existing database functions to call a new Database
object.  Once in place, the callers can be changed to call the new
Database methods directly.

Note that for many of the current functions, there is no functional
difference between the behavior of the tep_db_ and mysqli_ functions.
One of the steps in this project will be to get rid of most of the
places where the tep_db_ function is just a thin wrapper around a
mysqli_ function.

Also, the $$link shared global will be going away in favor of a
more traditional database object.  That will make more functions
simple wrappers (because a number only default the $link value).
This only leaves six method equivalents of the tep_db_ functions.

This currently just uses the mysqli_ functions or methods.  It is
possible that it would make more sense to switch to PDO.  Since
these changes are still several releases away, it is possible that
we may change directions and use PDO instead.  It is simpler to use
mysqli_ at the moment because that's what had been in use under the
hood.  There is still time to revisit that if there is some reason
to prefer PDO.